third_party/WebKit/LayoutTests/crypto/importKey-badParameters-expected.txt

CONSOLE ERROR: Invalid keyData argument
CONSOLE ERROR: Invalid keyData argument
CONSOLE ERROR: Algorithm: Not an object
CONSOLE ERROR: Invalid keyFormat argument
CONSOLE ERROR: Invalid keyUsages argument
CONSOLE ERROR: Invalid keyFormat argument
CONSOLE ERROR: Algorithm: HMAC: HmacImportParams: hash: Missing or not a dictionary
CONSOLE ERROR: Algorithm: SHA-1: Unsupported operation
Tests calling cypto.subtle.importKey with bad parameters

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS error is null
PASS error is null
PASS error is null
PASS error is null
PASS error is null
PASS error is null
PASS error is null
PASS error is null
PASS successfullyParsed is true

TEST COMPLETE
